+++ src/main/java/fr/gourmetgo/repository/RestaurantRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import fr.gourmetgo.entity.Restaurant;
 
-public interface RestaurantRepository extends JpaRepository <Restaurant, Integer>{
+public interface RestaurantRepository extends JpaRepository <Restaurant, Long>{
 
 }
